requirements.txt
# Core Django and Database
Django==4.2.7
psycopg2-binary==2.9.9
django-environ==0.11.2

# Authentication and User Management
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10

# Media Processing and Storage
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
ffmpeg-python==0.2.0

# AI and Machine Learning
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Background Tasks
celery==5.3.4
redis==5.0.1

# Search and Indexing
django-haystack==3.2.1
whoosh==2.7.4

# API and Web Services
requests==2.31.0
django-cors-headers==4.3.1

# Testing and Development
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0

# Security
django-csp==3.7
django-axes==6.3.0

# Monitoring and Logging
sentry-sdk==1.35.0
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
django-filter==23.3
django-widget-tweaks==1.5.0

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0 